+++ lib/src/implementation/states/extra_state.dart
@@ -120,6 +120,9 @@ class NetworkSearchableExtraState<T, E> extends NetworkSearchableState<T>
       query: query ?? this.query,
     );
   }
+
+  @override
+  List<Object?> get props => [...super.props, extraData];
 }
 
 class NetworkFilterableExtraState<T, F, E> extends NetworkFilterableState<T, F>
@@ -186,4 +189,7 @@ class NetworkFilterableExtraState<T, F, E> extends NetworkFilterableState<T, F>
       filter: filter ?? this.filter,
     );
   }
+
+  @override
+  List<Object?> get props => [...super.props, extraData];
 }